Entity Model Layer Reg, and move to new package
