+++ fabric/src/main/java/dev/architectury/registry/client/level/entity/fabric/EntityRendererRegistryImpl.java
@@ -17,9 +17,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package dev.architectury.registry.level.entity.fabric;
+package dev.architectury.registry.client.level.entity.fabric;
 
-import net.fabricmc.fabric.api.client.rendereregistry.v1.EntityRendererRegistry;
+import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
@@ -28,6 +28,6 @@ import java.util.function.Supplier;
 
 public class EntityRendererRegistryImpl {
     public static <T extends Entity> void register(Supplier<EntityType<? extends T>> type, EntityRendererProvider<T> provider) {
-        EntityRendererRegistry.INSTANCE.register(type.get(), provider);
+        EntityRendererRegistry.register(type.get(), provider);
     }
 }
